GenUi to cancel my Netflix
The prompt I gave for this was "I want to cancel my Netflix". This is something the user can say using their voice with Sky Remote.
I also starting giving the LLM placeholder images so now it has images!!
This again was generated by the LLM in under 2 minutes, no designer was needed to create the screen.